refactor(13-02): move featuregates + premium middleware into shared/
- Promote services/share-service/featuregates/ -> shared/featuregates/
  (byte-identical copy, test import path updated to shared/featuregates)
- Promote services/share-service/middleware/premium.go -> shared/middleware/premium.go
  (byte-identical copy; premium_test.go stays in package middleware)
- Update share-service cmd/main.go to import the promoted packages:
  featuregates now comes from shared, and the formerly aliased
  localMiddleware.RequirePremium is now middleware.RequirePremium (single
  shared/middleware import, no alias needed)
- Update services/share-service/handlers/admin_featuregates.go and
  admin_featuregates_test.go to reference shared/featuregates
- Remove the emptied services/share-service/{featuregates,middleware}
  directories and .gitkeep placeholder
- Run go mod tidy in both share-service and shared to regenerate graph

Why: Phase 13 needs overlay-manager to consume RequirePremium +
FeatureGateCache without duplicating code or creating a circular import
through share-service. Promoting to shared/ keeps the existing share-service
behavior (tests green) and unblocks the new TTS premium-gated endpoints.

Safety:
- No promauto imports in either moved file (verified) — shared libs must
  not register Prometheus metrics (ADR-0008 pitfall 6).
- grep -rn "caesar/all-chat/services/share-service" shared/ returns zero
  matches (no stale internal references).
- Existing share-service tests continue to pass; shared tests pass.

Addresses Phase 13 D-17.
