+++ services/share-service/handlers/admin_featuregates.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/caesar/all-chat/services/share-service/featuregates"
+	"github.com/caesar/all-chat/shared/featuregates"
 	"github.com/gin-gonic/gin"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/redis/go-redis/v9"
