+++ services/share-service/handlers/admin_featuregates_test.go
@@ -25,7 +25,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/caesar/all-chat/services/share-service/featuregates"
+	"github.com/caesar/all-chat/shared/featuregates"
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
